+++ src/updates.rs
@@ -1,6 +1,6 @@
 use crate::{
     block::BlockGenerator,
-    config::{Config, FileConfig, ThemeVars},
+    config::{BlockConfig, Config, FileConfig, ThemeVars},
     utils::expand_tilde,
 };
 use colored::Colorize;
@@ -19,12 +19,28 @@ pub fn run(theme_name: String, config: &Config) {
             exit(1);
         }
     };
-    let update = UpdatesGenerator::new(theme_name, theme.clone());
+    let update_gen = UpdatesGenerator::new(theme_name, theme.clone());
 
     for (_, conf) in &config.files {
-        if let Some(contents) = update.generate(conf) {
-            fs::write(expand_tilde(&conf.path), contents.as_bytes()).unwrap();
+        match conf {
+            FileConfig::Multi(mutli) => mutli.blocks.clone().into_iter().for_each(|(tag, c)| {
+                let config = BlockConfig {
+                    path: mutli.path.clone(),
+                    comment: mutli.comment.clone(),
+                    block: c,
+                };
+                write_update(&mutli.path, update_gen.generate(&config, Some(tag)));
+            }),
+            FileConfig::Single(config) => {
+                write_update(&config.path, update_gen.generate(config, None));
+            }
         }
+    }
+}
+
+fn write_update(path: &String, update: Option<String>) {
+    if let Some(update) = update {
+        fs::write(expand_tilde(&path), update.as_bytes()).unwrap();
     }
 }
 
@@ -38,24 +54,32 @@ impl UpdatesGenerator {
         Self { theme_name, theme }
     }
 
-    pub fn generate(&self, config: &FileConfig) -> Option<String> {
-        let contents = self.get_file_contents(config);
+    pub fn generate(&self, config: &BlockConfig, tag: Option<String>) -> Option<String> {
+        let blk_gen = BlockGenerator::new(
+            &self.theme_name,
+            &self.theme,
+            FileConfig::Single(config.clone()),
+        );
+        let tags = BlockGenerator::get_block_tags(&tag);
+
+        let contents = self.get_file_contents(config, &tags);
         if let Err(_) = contents {
             return None;
         }
         let contents = contents.unwrap();
 
-        let blk_gen = BlockGenerator::new(&self.theme_name, &self.theme, config);
         let mut new_block = blk_gen.generate();
         // Replacing dollar sign to avoid Regex issues
-        new_block = blk_gen.wrap(&new_block).replace("$", "$$");
+        new_block = blk_gen.wrap(&new_block, &tags).replace("$", "$$");
 
-        Some(Self::get_block_re(&config.comment)
-            .replacen(&contents, 1, new_block)
-            .to_string())
+        Some(
+            Self::get_block_re(&config.comment, &tags)
+                .replacen(&contents, 1, new_block)
+                .to_string(),
+        )
     }
 
-    fn get_file_contents(&self, conf: &FileConfig) -> Result<String, ()> {
+    fn get_file_contents(&self, conf: &BlockConfig, tags: &(String, String)) -> Result<String, ()> {
         let contents = match fs::read_to_string(expand_tilde(&conf.path)) {
             Ok(f) => f,
             Err(e) => {
@@ -64,17 +88,22 @@ impl UpdatesGenerator {
             }
         };
 
-        let themer_block_re = Self::get_block_re(&conf.comment);
+        let themer_block_re = Self::get_block_re(&conf.comment, tags);
         if !themer_block_re.is_match(&contents) {
-            log::warn!("Failed to find THEMER block inside of `{}`. Skipping it.", conf.path);
+            log::warn!(
+                "Failed to find THEMER block inside of `{}`. Skipping it.",
+                conf.path
+            );
             return Err(());
         }
 
         Ok(contents)
     }
 
-    pub fn get_block_re(comment: &String) -> Regex {
-        RegexBuilder::new(&format!("{0} THEMER\n.*{0} THEMER_END", comment))
+    pub fn get_block_re(comment: &String, (start, end): &(String, String)) -> Regex {
+        log::debug!("Generate regex for block `{comment} {start} ... {comment} {end}");
+
+        RegexBuilder::new(&format!("{0} {start}\n.*{0} {end}", comment))
             .dot_matches_new_line(true)
             .build()
             .unwrap()